singalingbackend/requirement.txt
flask-restful
flask-socketio
python-dotenv
Flask-And-Redis
eventlet
flask-cors
gunicorn
PyJWT
pymongo

